requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=10.0.0
faiss-cpu>=1.7.4
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
qwen-vl-utils>=0.0.2
